+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 All notable changes to this project will be documented in this file.
 
 # Unreleased
+- fix DuckDB::ScalarFunction NULL input handling.
+- fix DuckDB::ScalarFunction INTEGER output type (int32 vs int64).
 - add DuckDB::ScalarFunction#add_parameter.
 - add DuckDB::ScalarFunction#set_return_type.
 - bump bundler 4.0 in duckdb.gemspec.